// src/app/models/user.ts
export class User {

    _id: String ;
    nom:  String ;
    prenom: String ;
    poste:  String ;
    experience:  Number ;
    type_contrat: String ;
    date_entree:Date ;
    stack_principale: String ;
    stack_secondaire: String ;
    email :String;
    password:String;
    role:String;
  }